+++ Source/Details/ASPINRemoteImageDownloader.h
@@ -26,7 +26,6 @@ NS_ASSUME_NONNULL_BEGIN
  * This is the default downloader used by network backed image nodes if PINRemoteImage and PINCache are
  * available. It uses PINRemoteImage's features to provide caching and progressive image downloads.
  */
-@property (class, readonly) ASPINRemoteImageDownloader *sharedDownloader;
 + (ASPINRemoteImageDownloader *)sharedDownloader NS_RETURNS_RETAINED;
 
 
@@ -39,6 +38,16 @@ NS_ASSUME_NONNULL_BEGIN
  *
  */
 + (void)setSharedImageManagerWithConfiguration:(nullable NSURLSessionConfiguration *)configuration;
+
+/**
+ * Sets a custom preconfigured PINRemoteImageManager that will be used by @c ASNetworkImageNodes and @c ASMultiplexImageNodes
+ * while loading images off the network. This must be specified early in the application lifecycle before
+ * `sharedDownloader` is accessed. If nil is passed in as the PINRemoteImageManager, it will create
+ * a default image manager with a nil session configuration.
+ *
+ * @param preconfiguredPINRemoteImageManager The preconfigured remote image manager that will be used by `sharedDownloader`
+ */
++ (void)setSharedPreconfiguredRemoteImageManager:(nullable PINRemoteImageManager *)preconfiguredPINRemoteImageManager;
 
 /**
  * The shared instance of a @c PINRemoteImageManager used by all @c ASPINRemoteImageDownloaders